split configmap, downwardapi, and secrets, into two files each volume/non-volume and moved sig labels to front of the description string
