+++ test/e2e/common/BUILD
@@ -11,6 +11,7 @@ go_library(
         "apparmor.go",
         "autoscaling_utils.go",
         "configmap.go",
+        "configmap_volume.go",
         "container_probe.go",
         "docker_containers.go",
         "downward_api.go",
@@ -26,6 +27,7 @@ go_library(
         "privileged.go",
         "projected.go",
         "secrets.go",
+        "secrets_volume.go",
         "sysctl.go",
         "util.go",
         "volumes.go",
