+++ test/e2e/common/configmap.go
@@ -18,314 +18,16 @@ package common
 
 import (
 	"fmt"
-	"os"
-	"path"
 
 	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
 	"k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/util/uuid"
 	"k8s.io/kubernetes/test/e2e/framework"
 )
 
-var _ = framework.KubeDescribe("ConfigMap", func() {
+var _ = Describe("[sig-api-machinery] ConfigMap", func() {
 	f := framework.NewDefaultFramework("configmap")
-
-	It("should be consumable from pods in volume [Conformance] [sig-storage]", func() {
-		doConfigMapE2EWithoutMappings(f, 0, 0, nil)
-	})
-
-	It("should be consumable from pods in volume with defaultMode set [Conformance] [sig-storage]", func() {
-		defaultMode := int32(0400)
-		doConfigMapE2EWithoutMappings(f, 0, 0, &defaultMode)
-	})
-
-	It("should be consumable from pods in volume as non-root with defaultMode and fsGroup set [Feature:FSGroup] [sig-storage]", func() {
-		defaultMode := int32(0440) /* setting fsGroup sets mode to at least 440 */
-		doConfigMapE2EWithoutMappings(f, 1000, 1001, &defaultMode)
-	})
-
-	It("should be consumable from pods in volume as non-root [Conformance] [sig-storage]", func() {
-		doConfigMapE2EWithoutMappings(f, 1000, 0, nil)
-	})
-
-	It("should be consumable from pods in volume as non-root with FSGroup [Feature:FSGroup] [sig-storage]", func() {
-		doConfigMapE2EWithoutMappings(f, 1000, 1001, nil)
-	})
-
-	It("should be consumable from pods in volume with mappings [Conformance] [sig-storage]", func() {
-		doConfigMapE2EWithMappings(f, 0, 0, nil)
-	})
-
-	It("should be consumable from pods in volume with mappings and Item mode set[Conformance] [sig-storage]", func() {
-		mode := int32(0400)
-		doConfigMapE2EWithMappings(f, 0, 0, &mode)
-	})
-
-	It("should be consumable from pods in volume with mappings as non-root [Conformance] [sig-storage]", func() {
-		doConfigMapE2EWithMappings(f, 1000, 0, nil)
-	})
-
-	It("should be consumable from pods in volume with mappings as non-root with FSGroup [Feature:FSGroup] [sig-storage]", func() {
-		doConfigMapE2EWithMappings(f, 1000, 1001, nil)
-	})
-
-	It("updates should be reflected in volume [Conformance] [sig-storage]", func() {
-		podLogTimeout := framework.GetPodSecretUpdateTimeout(f.ClientSet)
-		containerTimeoutArg := fmt.Sprintf("--retry_time=%v", int(podLogTimeout.Seconds()))
-
-		name := "configmap-test-upd-" + string(uuid.NewUUID())
-		volumeName := "configmap-volume"
-		volumeMountPath := "/etc/configmap-volume"
-		containerName := "configmap-volume-test"
-
-		configMap := &v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: f.Namespace.Name,
-				Name:      name,
-			},
-			Data: map[string]string{
-				"data-1": "value-1",
-			},
-		}
-
-		By(fmt.Sprintf("Creating configMap with name %s", configMap.Name))
-		var err error
-		if configMap, err = f.ClientSet.Core().ConfigMaps(f.Namespace.Name).Create(configMap); err != nil {
-			framework.Failf("unable to create test configMap %s: %v", configMap.Name, err)
-		}
-
-		pod := &v1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "pod-configmaps-" + string(uuid.NewUUID()),
-			},
-			Spec: v1.PodSpec{
-				Volumes: []v1.Volume{
-					{
-						Name: volumeName,
-						VolumeSource: v1.VolumeSource{
-							ConfigMap: &v1.ConfigMapVolumeSource{
-								LocalObjectReference: v1.LocalObjectReference{
-									Name: name,
-								},
-							},
-						},
-					},
-				},
-				Containers: []v1.Container{
-					{
-						Name:    containerName,
-						Image:   mountImage,
-						Command: []string{"/mounttest", "--break_on_expected_content=false", containerTimeoutArg, "--file_content_in_loop=/etc/configmap-volume/data-1"},
-						VolumeMounts: []v1.VolumeMount{
-							{
-								Name:      volumeName,
-								MountPath: volumeMountPath,
-								ReadOnly:  true,
-							},
-						},
-					},
-				},
-				RestartPolicy: v1.RestartPolicyNever,
-			},
-		}
-		By("Creating the pod")
-		f.PodClient().CreateSync(pod)
-
-		pollLogs := func() (string, error) {
-			return framework.GetPodLogs(f.ClientSet, f.Namespace.Name, pod.Name, containerName)
-		}
-
-		Eventually(pollLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("value-1"))
-
-		By(fmt.Sprintf("Updating configmap %v", configMap.Name))
-		configMap.ResourceVersion = "" // to force update
-		configMap.Data["data-1"] = "value-2"
-		_, err = f.ClientSet.Core().ConfigMaps(f.Namespace.Name).Update(configMap)
-		Expect(err).NotTo(HaveOccurred(), "Failed to update configmap %q in namespace %q", configMap.Name, f.Namespace.Name)
-
-		By("waiting to observe update in volume")
-		Eventually(pollLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("value-2"))
-	})
-
-	It("optional updates should be reflected in volume [Conformance] [sig-storage]", func() {
-		podLogTimeout := framework.GetPodSecretUpdateTimeout(f.ClientSet)
-		containerTimeoutArg := fmt.Sprintf("--retry_time=%v", int(podLogTimeout.Seconds()))
-		trueVal := true
-		volumeMountPath := "/etc/configmap-volumes"
-
-		deleteName := "cm-test-opt-del-" + string(uuid.NewUUID())
-		deleteContainerName := "delcm-volume-test"
-		deleteVolumeName := "deletecm-volume"
-		deleteConfigMap := &v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: f.Namespace.Name,
-				Name:      deleteName,
-			},
-			Data: map[string]string{
-				"data-1": "value-1",
-			},
-		}
-
-		updateName := "cm-test-opt-upd-" + string(uuid.NewUUID())
-		updateContainerName := "updcm-volume-test"
-		updateVolumeName := "updatecm-volume"
-		updateConfigMap := &v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: f.Namespace.Name,
-				Name:      updateName,
-			},
-			Data: map[string]string{
-				"data-1": "value-1",
-			},
-		}
-
-		createName := "cm-test-opt-create-" + string(uuid.NewUUID())
-		createContainerName := "createcm-volume-test"
-		createVolumeName := "createcm-volume"
-		createConfigMap := &v1.ConfigMap{
-			ObjectMeta: metav1.ObjectMeta{
-				Namespace: f.Namespace.Name,
-				Name:      createName,
-			},
-			Data: map[string]string{
-				"data-1": "value-1",
-			},
-		}
-
-		By(fmt.Sprintf("Creating configMap with name %s", deleteConfigMap.Name))
-		var err error
-		if deleteConfigMap, err = f.ClientSet.Core().ConfigMaps(f.Namespace.Name).Create(deleteConfigMap); err != nil {
-			framework.Failf("unable to create test configMap %s: %v", deleteConfigMap.Name, err)
-		}
-
-		By(fmt.Sprintf("Creating configMap with name %s", updateConfigMap.Name))
-		if updateConfigMap, err = f.ClientSet.Core().ConfigMaps(f.Namespace.Name).Create(updateConfigMap); err != nil {
-			framework.Failf("unable to create test configMap %s: %v", updateConfigMap.Name, err)
-		}
-
-		pod := &v1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "pod-configmaps-" + string(uuid.NewUUID()),
-			},
-			Spec: v1.PodSpec{
-				Volumes: []v1.Volume{
-					{
-						Name: deleteVolumeName,
-						VolumeSource: v1.VolumeSource{
-							ConfigMap: &v1.ConfigMapVolumeSource{
-								LocalObjectReference: v1.LocalObjectReference{
-									Name: deleteName,
-								},
-								Optional: &trueVal,
-							},
-						},
-					},
-					{
-						Name: updateVolumeName,
-						VolumeSource: v1.VolumeSource{
-							ConfigMap: &v1.ConfigMapVolumeSource{
-								LocalObjectReference: v1.LocalObjectReference{
-									Name: updateName,
-								},
-								Optional: &trueVal,
-							},
-						},
-					},
-					{
-						Name: createVolumeName,
-						VolumeSource: v1.VolumeSource{
-							ConfigMap: &v1.ConfigMapVolumeSource{
-								LocalObjectReference: v1.LocalObjectReference{
-									Name: createName,
-								},
-								Optional: &trueVal,
-							},
-						},
-					},
-				},
-				Containers: []v1.Container{
-					{
-						Name:    deleteContainerName,
-						Image:   mountImage,
-						Command: []string{"/mounttest", "--break_on_expected_content=false", containerTimeoutArg, "--file_content_in_loop=/etc/configmap-volumes/delete/data-1"},
-						VolumeMounts: []v1.VolumeMount{
-							{
-								Name:      deleteVolumeName,
-								MountPath: path.Join(volumeMountPath, "delete"),
-								ReadOnly:  true,
-							},
-						},
-					},
-					{
-						Name:    updateContainerName,
-						Image:   mountImage,
-						Command: []string{"/mounttest", "--break_on_expected_content=false", containerTimeoutArg, "--file_content_in_loop=/etc/configmap-volumes/update/data-3"},
-						VolumeMounts: []v1.VolumeMount{
-							{
-								Name:      updateVolumeName,
-								MountPath: path.Join(volumeMountPath, "update"),
-								ReadOnly:  true,
-							},
-						},
-					},
-					{
-						Name:    createContainerName,
-						Image:   mountImage,
-						Command: []string{"/mounttest", "--break_on_expected_content=false", containerTimeoutArg, "--file_content_in_loop=/etc/configmap-volumes/create/data-1"},
-						VolumeMounts: []v1.VolumeMount{
-							{
-								Name:      createVolumeName,
-								MountPath: path.Join(volumeMountPath, "create"),
-								ReadOnly:  true,
-							},
-						},
-					},
-				},
-				RestartPolicy: v1.RestartPolicyNever,
-			},
-		}
-		By("Creating the pod")
-		f.PodClient().CreateSync(pod)
-
-		pollCreateLogs := func() (string, error) {
-			return framework.GetPodLogs(f.ClientSet, f.Namespace.Name, pod.Name, createContainerName)
-		}
-		Eventually(pollCreateLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("Error reading file /etc/configmap-volumes/create/data-1"))
-
-		pollUpdateLogs := func() (string, error) {
-			return framework.GetPodLogs(f.ClientSet, f.Namespace.Name, pod.Name, updateContainerName)
-		}
-		Eventually(pollUpdateLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("Error reading file /etc/configmap-volumes/update/data-3"))
-
-		pollDeleteLogs := func() (string, error) {
-			return framework.GetPodLogs(f.ClientSet, f.Namespace.Name, pod.Name, deleteContainerName)
-		}
-		Eventually(pollDeleteLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("value-1"))
-
-		By(fmt.Sprintf("Deleting configmap %v", deleteConfigMap.Name))
-		err = f.ClientSet.Core().ConfigMaps(f.Namespace.Name).Delete(deleteConfigMap.Name, &metav1.DeleteOptions{})
-		Expect(err).NotTo(HaveOccurred(), "Failed to delete configmap %q in namespace %q", deleteConfigMap.Name, f.Namespace.Name)
-
-		By(fmt.Sprintf("Updating configmap %v", updateConfigMap.Name))
-		updateConfigMap.ResourceVersion = "" // to force update
-		delete(updateConfigMap.Data, "data-1")
-		updateConfigMap.Data["data-3"] = "value-3"
-		_, err = f.ClientSet.Core().ConfigMaps(f.Namespace.Name).Update(updateConfigMap)
-		Expect(err).NotTo(HaveOccurred(), "Failed to update configmap %q in namespace %q", updateConfigMap.Name, f.Namespace.Name)
-
-		By(fmt.Sprintf("Creating configMap with name %s", createConfigMap.Name))
-		if createConfigMap, err = f.ClientSet.Core().ConfigMaps(f.Namespace.Name).Create(createConfigMap); err != nil {
-			framework.Failf("unable to create test configMap %s: %v", createConfigMap.Name, err)
-		}
-
-		By("waiting to observe update in volume")
-
-		Eventually(pollCreateLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("value-1"))
-		Eventually(pollUpdateLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("value-3"))
-		Eventually(pollDeleteLogs, podLogTimeout, framework.Poll).Should(ContainSubstring("Error reading file /etc/configmap-volumes/delete/data-1"))
-	})
 
 	It("should be consumable via environment variable [Conformance]", func() {
 		name := "configmap-test-" + string(uuid.NewUUID())
@@ -409,93 +111,7 @@ var _ = framework.KubeDescribe("ConfigMap", func() {
 			"p_data_1=value-1", "p_data_2=value-2", "p_data_3=value-3",
 		})
 	})
-
-	It("should be consumable in multiple volumes in the same pod [Conformance] [sig-storage]", func() {
-		var (
-			name             = "configmap-test-volume-" + string(uuid.NewUUID())
-			volumeName       = "configmap-volume"
-			volumeMountPath  = "/etc/configmap-volume"
-			volumeName2      = "configmap-volume-2"
-			volumeMountPath2 = "/etc/configmap-volume-2"
-			configMap        = newConfigMap(f, name)
-		)
-
-		By(fmt.Sprintf("Creating configMap with name %s", configMap.Name))
-		var err error
-		if configMap, err = f.ClientSet.Core().ConfigMaps(f.Namespace.Name).Create(configMap); err != nil {
-			framework.Failf("unable to create test configMap %s: %v", configMap.Name, err)
-		}
-
-		pod := &v1.Pod{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "pod-configmaps-" + string(uuid.NewUUID()),
-			},
-			Spec: v1.PodSpec{
-				Volumes: []v1.Volume{
-					{
-						Name: volumeName,
-						VolumeSource: v1.VolumeSource{
-							ConfigMap: &v1.ConfigMapVolumeSource{
-								LocalObjectReference: v1.LocalObjectReference{
-									Name: name,
-								},
-							},
-						},
-					},
-					{
-						Name: volumeName2,
-						VolumeSource: v1.VolumeSource{
-							ConfigMap: &v1.ConfigMapVolumeSource{
-								LocalObjectReference: v1.LocalObjectReference{
-									Name: name,
-								},
-							},
-						},
-					},
-				},
-				Containers: []v1.Container{
-					{
-						Name:  "configmap-volume-test",
-						Image: mountImage,
-						Args:  []string{"--file_content=/etc/configmap-volume/data-1"},
-						VolumeMounts: []v1.VolumeMount{
-							{
-								Name:      volumeName,
-								MountPath: volumeMountPath,
-								ReadOnly:  true,
-							},
-							{
-								Name:      volumeName2,
-								MountPath: volumeMountPath2,
-								ReadOnly:  true,
-							},
-						},
-					},
-				},
-				RestartPolicy: v1.RestartPolicyNever,
-			},
-		}
-
-		f.TestContainerOutput("consume configMaps", pod, 0, []string{
-			"content of file \"/etc/configmap-volume/data-1\": value-1",
-		})
-
-	})
 })
-
-func newConfigMap(f *framework.Framework, name string) *v1.ConfigMap {
-	return &v1.ConfigMap{
-		ObjectMeta: metav1.ObjectMeta{
-			Namespace: f.Namespace.Name,
-			Name:      name,
-		},
-		Data: map[string]string{
-			"data-1": "value-1",
-			"data-2": "value-2",
-			"data-3": "value-3",
-		},
-	}
-}
 
 func newEnvFromConfigMap(f *framework.Framework, name string) *v1.ConfigMap {
 	return &v1.ConfigMap{
@@ -509,173 +125,4 @@ func newEnvFromConfigMap(f *framework.Framework, name string) *v1.ConfigMap {
 			"data_3": "value-3",
 		},
 	}
-}
-
-func doConfigMapE2EWithoutMappings(f *framework.Framework, uid, fsGroup int64, defaultMode *int32) {
-	userID := int64(uid)
-	groupID := int64(fsGroup)
-
-	var (
-		name            = "configmap-test-volume-" + string(uuid.NewUUID())
-		volumeName      = "configmap-volume"
-		volumeMountPath = "/etc/configmap-volume"
-		configMap       = newConfigMap(f, name)
-	)
-
-	By(fmt.Sprintf("Creating configMap with name %s", configMap.Name))
-	var err error
-	if configMap, err = f.ClientSet.Core().ConfigMaps(f.Namespace.Name).Create(configMap); err != nil {
-		framework.Failf("unable to create test configMap %s: %v", configMap.Name, err)
-	}
-
-	one := int64(1)
-	pod := &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "pod-configmaps-" + string(uuid.NewUUID()),
-		},
-		Spec: v1.PodSpec{
-			SecurityContext: &v1.PodSecurityContext{},
-			Volumes: []v1.Volume{
-				{
-					Name: volumeName,
-					VolumeSource: v1.VolumeSource{
-						ConfigMap: &v1.ConfigMapVolumeSource{
-							LocalObjectReference: v1.LocalObjectReference{
-								Name: name,
-							},
-						},
-					},
-				},
-			},
-			Containers: []v1.Container{
-				{
-					Name:  "configmap-volume-test",
-					Image: mountImage,
-					Args: []string{
-						"--file_content=/etc/configmap-volume/data-1",
-						"--file_mode=/etc/configmap-volume/data-1"},
-					VolumeMounts: []v1.VolumeMount{
-						{
-							Name:      volumeName,
-							MountPath: volumeMountPath,
-						},
-					},
-				},
-			},
-			RestartPolicy:                 v1.RestartPolicyNever,
-			TerminationGracePeriodSeconds: &one,
-		},
-	}
-
-	if userID != 0 {
-		pod.Spec.SecurityContext.RunAsUser = &userID
-	}
-
-	if groupID != 0 {
-		pod.Spec.SecurityContext.FSGroup = &groupID
-	}
-
-	if defaultMode != nil {
-		pod.Spec.Volumes[0].VolumeSource.ConfigMap.DefaultMode = defaultMode
-	} else {
-		mode := int32(0644)
-		defaultMode = &mode
-	}
-
-	modeString := fmt.Sprintf("%v", os.FileMode(*defaultMode))
-	output := []string{
-		"content of file \"/etc/configmap-volume/data-1\": value-1",
-		"mode of file \"/etc/configmap-volume/data-1\": " + modeString,
-	}
-	f.TestContainerOutput("consume configMaps", pod, 0, output)
-}
-
-func doConfigMapE2EWithMappings(f *framework.Framework, uid, fsGroup int64, itemMode *int32) {
-	userID := int64(uid)
-	groupID := int64(fsGroup)
-
-	var (
-		name            = "configmap-test-volume-map-" + string(uuid.NewUUID())
-		volumeName      = "configmap-volume"
-		volumeMountPath = "/etc/configmap-volume"
-		configMap       = newConfigMap(f, name)
-	)
-
-	By(fmt.Sprintf("Creating configMap with name %s", configMap.Name))
-
-	var err error
-	if configMap, err = f.ClientSet.Core().ConfigMaps(f.Namespace.Name).Create(configMap); err != nil {
-		framework.Failf("unable to create test configMap %s: %v", configMap.Name, err)
-	}
-
-	one := int64(1)
-	pod := &v1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: "pod-configmaps-" + string(uuid.NewUUID()),
-		},
-		Spec: v1.PodSpec{
-			SecurityContext: &v1.PodSecurityContext{},
-			Volumes: []v1.Volume{
-				{
-					Name: volumeName,
-					VolumeSource: v1.VolumeSource{
-						ConfigMap: &v1.ConfigMapVolumeSource{
-							LocalObjectReference: v1.LocalObjectReference{
-								Name: name,
-							},
-							Items: []v1.KeyToPath{
-								{
-									Key:  "data-2",
-									Path: "path/to/data-2",
-								},
-							},
-						},
-					},
-				},
-			},
-			Containers: []v1.Container{
-				{
-					Name:  "configmap-volume-test",
-					Image: mountImage,
-					Args: []string{"--file_content=/etc/configmap-volume/path/to/data-2",
-						"--file_mode=/etc/configmap-volume/path/to/data-2"},
-					VolumeMounts: []v1.VolumeMount{
-						{
-							Name:      volumeName,
-							MountPath: volumeMountPath,
-							ReadOnly:  true,
-						},
-					},
-				},
-			},
-			RestartPolicy:                 v1.RestartPolicyNever,
-			TerminationGracePeriodSeconds: &one,
-		},
-	}
-
-	if userID != 0 {
-		pod.Spec.SecurityContext.RunAsUser = &userID
-	}
-
-	if groupID != 0 {
-		pod.Spec.SecurityContext.FSGroup = &groupID
-	}
-
-	if itemMode != nil {
-		pod.Spec.Volumes[0].VolumeSource.ConfigMap.Items[0].Mode = itemMode
-	} else {
-		mode := int32(0644)
-		itemMode = &mode
-	}
-
-	// Just check file mode if fsGroup is not set. If fsGroup is set, the
-	// final mode is adjusted and we are not testing that case.
-	output := []string{
-		"content of file \"/etc/configmap-volume/path/to/data-2\": value-2",
-	}
-	if fsGroup == 0 {
-		modeString := fmt.Sprintf("%v", os.FileMode(*itemMode))
-		output = append(output, "mode of file \"/etc/configmap-volume/path/to/data-2\": "+modeString)
-	}
-	f.TestContainerOutput("consume configMaps", pod, 0, output)
 }
